+++ backend/wmg/data/transform.py
@@ -1,6 +1,18 @@
+import gc
+import time
 from typing import Dict
+
+import anndata
+import numpy
+import numpy as np
 import tiledb
 import pandas as pd
+from scipy import sparse
+from scipy.sparse import csr_matrix, coo_matrix
+
+from backend.wmg.data.load_corpus import get_X_raw, logger
+from backend.wmg.data.wmg_constants import RANKIT_RAW_EXPR_COUNT_FILTERING_MIN_THRESHOLD, INTEGRATED_ARRAY_NAME
+from backend.wmg.data.rankit import rankit
 
 from backend.wmg.data.snapshot import CELL_TYPE_ORDERINGS_FILENAME
 
@@ -13,9 +25,9 @@ def get_cell_types_by_tissue(corpus_group: str) -> Dict:
     with tiledb.open(f"{corpus_group}/obs", "r") as obs:
         tissue_cell_types = (
             obs.query(attrs=[], dims=["tissue_ontology_term_id", "cell_type_ontology_term_id"])
-            .df[:]
-            .drop_duplicates()
-            .sort_values(by="tissue_ontology_term_id")
+                .df[:]
+                .drop_duplicates()
+                .sort_values(by="tissue_ontology_term_id")
         )
     unique_tissue_ontology_term_id = tissue_cell_types.tissue_ontology_term_id.unique()
     cell_type_by_tissue = {}
@@ -27,7 +39,7 @@ def get_cell_types_by_tissue(corpus_group: str) -> Dict:
     return cell_type_by_tissue
 
 
-def generate_cell_ordering(snapshot_path: str, cell_type_by_tissue: Dict) -> None:
+def generate_cell_ordering(snapshot_path: str, corpus_path: str) -> None:
     """
     Use graphviz to map all the cells associated with a tissue to the ontology tree and return their correct order
     """
@@ -37,7 +49,7 @@ def generate_cell_ordering(snapshot_path: str, cell_type_by_tissue: Dict) -> Non
     import pygraphviz as pgv
 
     onto = Ontology.from_obo_library("cl-basic.obo")
-
+    cell_type_by_tissue = get_cell_types_by_tissue(corpus_path)
     def compute_ordering(cells, root):
         ancestors = [list(onto[t].superclasses()) for t in cells if t in onto]
         ancestors = [i for s in ancestors for i in s]
@@ -84,3 +96,100 @@ def generate_cell_ordering(snapshot_path: str, cell_type_by_tissue: Dict) -> Non
 
     df = pd.DataFrame(data, columns=["tissue_ontology_term_id", "cell_type_ontology_term_id", "order"])
     df.to_json(f"{snapshot_path}/{CELL_TYPE_ORDERINGS_FILENAME}")
+
+
+
+# Filter cells
+def filter_out_rankits_with_low_expression_counts(
+        rankits: csr_matrix, raw_counts_csr: csr_matrix, expect_majority_filtered=True
+) -> coo_matrix:
+    """
+    Keep only rankit values that were computed from expression values above the desired threshold.
+
+    @param expect_majority_filtered: Set this to True if the caller expects the majority of rankit values will be
+    filtered, as we can then use an optimal implementation
+    """
+
+    rankits_nnz_orig = rankits.nnz
+    raw_counts = raw_counts_csr.tocoo(copy=False)
+    to_keep_mask = raw_counts.data > RANKIT_RAW_EXPR_COUNT_FILTERING_MIN_THRESHOLD
+
+    # Unfortunately, it seems to take longer to decide which strategy to use than it does to just run either one.
+    # start = time.time()
+    # expect_majority_filtered = to_keep_mask.sum() < rankits_nnz_orig / 2
+    # end = time.time()
+    # logger.info(f"filter_out_rankits_with_low_expression_counts(): use extract strategy={expect_majority_filtered}; "
+    #             f"decision time={end - start}")
+
+    start = time.time()
+
+    if expect_majority_filtered:
+        rows_to_keep = raw_counts.row[to_keep_mask]
+        cols_to_keep = raw_counts.col[to_keep_mask]
+        rankits_to_keep = rankits.data[to_keep_mask]
+        rankits_filtered = coo_matrix((rankits_to_keep, (rows_to_keep, cols_to_keep)))
+    else:
+        to_filter_mask = ~to_keep_mask
+        rows_to_filter = raw_counts.row[to_filter_mask]
+        cols_to_filter = raw_counts.col[to_filter_mask]
+        rankits[rows_to_filter, cols_to_filter] = 0.0
+        rankits.eliminate_zeros()
+        rankits_filtered = rankits.tocoo()
+
+    end = time.time()
+
+    logger.info(
+        f"filter duration={end - start}, "
+        f"orig size={rankits_nnz_orig}, "
+        f"abs reduction={rankits_nnz_orig - rankits_filtered.nnz}, "
+        f"% reduction={(rankits_nnz_orig - rankits_filtered.nnz) / rankits_nnz_orig}"
+    )
+
+    return rankits_filtered
+
+
+# Transform the expression matrix
+def transform_expression_raw_counts_to_rankit(
+        anndata_object: anndata.AnnData, corpus_path: str, global_var_index: numpy.ndarray, first_obs_idx: int
+):
+    """
+    Apply rankit normalization to raw count expression values and save to the tiledb corpus object
+    """
+    array_name = f"{corpus_path}/{INTEGRATED_ARRAY_NAME}"
+    expression_matrix = get_X_raw(anndata_object)
+    logger.info(f"saving {array_name}...")
+    stride = max(int(np.power(10, np.around(np.log10(1e9 / expression_matrix.shape[1])))), 10_000)
+    with tiledb.open(array_name, mode="w") as array:
+        for start in range(0, expression_matrix.shape[0], stride):
+            end = min(start + stride, expression_matrix.shape[0])
+            raw_expression_csr_matrix = sparse.csr_matrix(expression_matrix[start:end, :])
+
+            # Compute RankIt
+            rankit_integrated_csr_matrix = rankit(raw_expression_csr_matrix)
+
+            rankit_integrated_coo_matrix = filter_out_rankits_with_low_expression_counts(
+                rankit_integrated_csr_matrix, raw_expression_csr_matrix, expect_majority_filtered=True
+            )
+
+            global_rows = rankit_integrated_coo_matrix.row + start + first_obs_idx
+            global_cols = global_var_index[rankit_integrated_coo_matrix.col]
+
+            rankit_data = rankit_integrated_coo_matrix.data
+
+            assert len(rankit_data) == len(global_rows)
+            assert len(rankit_data) == len(global_cols)
+
+            array[global_rows, global_cols] = {"rankit": rankit_data}
+            del (
+                raw_expression_csr_matrix,
+                rankit_integrated_coo_matrix,
+                rankit_integrated_csr_matrix,
+                global_rows,
+                global_cols,
+                rankit_data,
+            )
+            gc.collect()
+
+    logger.debug(f"Saved {array_name}.")
+
+
